+++ js/share.js
@@ -1,30 +1,34 @@
-( function( window, document, undefined ) {
+(function() {
 
-    document.addEventListener( "DOMContentLoaded", function() {
+    var sharer;
 
-        var shareLinks = document.getElementsByClassName( 'share__link' ),
-            sharer;
+    // to make sure we always trigger a share links click (also on ajax loaded ones)
+    // we listen to the document click and work our way up the levels
+    document.addEventListener( 'click', function( e ) {
+        // clicked element
+        var $link = e.target;
 
-        if ( !shareLinks.length ) return;
-
-        for ( var i = 0; i < shareLinks.length; i++ ) {
-            if ( shareLinks[i].classList.contains( 'share__comments' ) ) continue;
-
-            shareLinks[i].addEventListener( 'click', function( e ) {
-                if ( this.href.match( new RegExp( '^(mailto|javascript|whatsapp|sms):' ) ) ) return;
-
-                e.preventDefault();
-
-                // Only allow ONE sharer popup
-                if (typeof sharer != 'undefined') {
-                    sharer.close();
-                    delete sharer;
-                }
-
-                sharer = window.open( this.href , 'sharer', 'height=440, width=640, toolbar=no, menubar=no, scrollbars=no, resizable=no, location=no, directories=no, status=no');
-            } );
+        // try to get the share link
+        while ( $link.tagName != 'A' || $link.className.split( ' ' ).indexOf( 'share__link' ) < 0 ) {
+            if ( $link instanceof HTMLBodyElement || !$link.parentElement ) return;
+            $link = $link.parentElement;
         }
 
+        // now we are pretty sure the user clicked on a share link
+
+        // no sharer popup for ...
+        if ( $link.href.match( new RegExp( '^(mailto|javascript|whatsapp|sms):' ) ) ) return;
+
+        e.preventDefault();
+
+        // only ONE sharer popup
+        if (typeof sharer != 'undefined') {
+            sharer.close();
+            delete sharer;
+        }
+
+        // eventually open sharer popup
+        sharer = window.open( $link.href , 'sharer', 'height=440, width=640, toolbar=no, menubar=no, scrollbars=no, resizable=no, location=no, directories=no, status=no');
     }, false );
 
-} )( this, this.document );
+})();