+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 **Fixed**
 
 * Get $post by URL.
+* Link click listener on EVERY (also ajax loaded) share link.
 
 ## 1.5.2 - 2018-10-05
 **Fixed**